requirements.txt
google-api-python-client==2.107.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
tqdm==4.66.1
pandas==2.2.0
tabulate==0.9.0
